Assemble.c: changed type in symbol_table to be uint16_t to match the address format in J type instructions

+++ src/assemble.c
@@ -3,7 +3,7 @@
 
 struct symbol_table {
     char *symbol;
-    int  address;
+    uint16_t  address;
     struct symbol_table *next;
 };
